Added bulk insert methods for Edgar data, and several bugfixes
Other minor changes/fixes include:
- Updated namespaces to `Stocks.DataModels` for better organization.
- Updated SQL schema to use `bigint` for `unit_id` and added new indexes.

+++ src/DataModels/XbrlFileModels/XbrlJson.cs
@@ -4,7 +4,7 @@ namespace DataModels.XbrlFileModels;
 
 public record XbrlJson
 {
-    [JsonPropertyName("cik")] public int Cik { get; init; }
+    [JsonPropertyName("cik")] public ulong Cik { get; init; }
     [JsonPropertyName("entityName")] public string EntityName { get; init; } = "";
     [JsonPropertyName("facts")] public Facts Facts { get; init; } = new();
